R 2.0.6 @ 2019-10-22

+++ SPX_libs/l_rtc79410.h
@@ -33,7 +33,7 @@ typedef struct
 	// date
 	uint8_t day;
 	uint8_t month;
-	uint16_t year;
+	uint8_t year;
 
 } RtcTimeType_t;
